requirements.txt
Django==2.1.5
django-crispy-forms==1.7.2
django-reversion==3.0.3
django-versatileimagefield==1.10
djangorestframework==3.9.1
djangorestframework-datatables==0.4.1
Pillow==5.4.1
pytz==2018.9
toml==0.10.0
